refactor(storage): extract write value serializer
